+++ tests/testthat/test-enw_model.R
@@ -1,14 +1,14 @@
 test_that("enw_model can compile the default model", {
   skip_on_cran()
-  enw_model()
+  expect_error(enw_model(), NA)
 })
 
 test_that("enw_model can compile the multi-threaded model", {
   skip_on_cran()
-  enw_model(threads = TRUE)
+  expect_error(enw_model(threads = TRUE), NA)
 })
 
 test_that("enw_model can compile using profiling", {
   skip_on_cran()
-  enw_model(profile = TRUE)
+  expect_error(enw_model(profile = TRUE), NA)
 })